Fix bug monitor publish lookup regressions
